base implementation for Buildings, Floors and Rooms
Started implementing the very base of the application: the ability to visualize Buildings, Floors and Rooms.
As of now, there isn't a dedicated interface to create instances of Buildings, Floors and Rooms (apart from Django Admin of course).
To create instances of those classes, it is necessary to use either:
- The Django Admin application
- The Django shell

This basic implementation also include the base template (built with tailwindcss + Flowbite) that will be used for all future pages.