added pictures for NDC and NDL
In order to have its image displayed, a Building must have an associated JPG image with the name of the building as its filename.
For instance, NDC will look for NDC.jpg in the static images directory.
This is not ideal (as indicated in the TODO in the template) as the images can't be edited though the application and the file extension is restricted to JPG.