updated models: Building, Floor, Room
Added a new "picture" attribute which contains the filename of the picture.
The Building.name attribute has been made unique so that it will be usable in URLs instead of the Building id.
The Room.number attribute has been made unique so that it will be usable in URLs instead of the Room's id.