updated URLs naming scheme: replaced Floor.id with Floor.number
A unique_together constraint has been added to the Floor model to ensure that
a Building cannot have two (or more) Floors that share the same number. That ensures that
the floor number is usable in the URL relative to the (unique) building name.

That means that we can use the Floor number (relative to the Building name) in URLs instead of the less meaningful id.
For instance, the URLs below now convey more meaning that before:
- aubouleau.fr/NDL/floors/11 -> aubouleau.fr/NDL/floors/3 (actual 3rd floor)
- aubouleau.fr/NDL/floors/12 -> aubouleau.fr/NDL/floors/4 (actual 4th floor)

For instance, the URLs below redirect to different floors although the floor number is the same
because the floor number is RELATIVE to the building name shown in the URL:
- aubouleau.fr/NDC/floors/2 -> redirects to the second floor of NDC
- aubouleau.fr/NDL/floors/2 -> redirects to the second floor of NDL