fixed HTTP 500 on the /buildings endpoint
Since 45d3c01e7afad8192200414e3c0f0b072f9b5355, the application is under the root URL / instead of /aubouleau.
This caused the automatic favicon fetching (/favicon.ico) to look for a building named "favicon.ico" since the first
argument after the root URL / is supposed to be a building name.
The view now correctly returns an HTTP 404 when the requested building does not exist.