added auto-creation of time slots
The time slots of a room (understand, the differents events that make up the planning of a room) are now created automatically.
A cronjob is setup in settings.py to run the TimeSlot.update_time_slots() static method every 6 hours.
This method launches the hpfetch.py script which downloads the calendars of all the different rooms from Hyperplanning.
Each calendar is then parsed to extract all the events for the current day. For each event, the following data is fetched:
- The datetime of the start of the event
- The datatime of the end of the event
- The summary (name) of the event

This works in the docker deployment.