added the room schedule on the room detail page
The room detail page now displays all the time slots that make up the schedule of the room.
Note that the schedule, as of now, only displas time slots for the current day.
The day has been set to begin at 08:00 and to end at 20:00. Whenever there are no time slots during
certain times, the schedule displays time slots marked as available.

Note that time slots can begin before the start of the day as well as end after the end of the day.
Those arbitrary values have been chosen to delimit the time at which the schedule will not be "filled up" with
time slots marked as available.
For instance, if the last time slot of the day ends at 18:00, a time slot marked as available will be created
from 18:00 to 20:00 (which is the end of the day). This works the same way for the beginning of the day.