added administration for buildings
There is now an administration part that allows a privileged user to create, edit and delete buildings.
In order to do so, the user must be in the "Administrators" group that is created upon startup of the application.
As of now, users can only be added to this group through the Django Admin app using the default superuser.
Users that are in the "Administrators" group are able to see a dedicated "Administration" menu on the sidebar, that
provides access to the current and future administration pages. A custom DTL filter has been added to check wheter or
not the authenticated user belongs to the "Administrators" group.

Additionally, users with insufficient privileges will receive HTTP 403 reponses when accessing the administration pages.

Note that, this commit also includes the following important changes:
- The logo has been changed with the definitive one
- The favicon has been added
- The docker-compose.yml file has been updated to fix the following issues:
  - Submitting forms no longer fails due to invalid CSRF validation. The new CSRF_TRUSTED_ORIGIN environment variable must
    now contain the list of domains allowed to submit HTTP requests.
  - The Django Admin app static files now load correctly (previously, the Django Admin app static files were not collected
    and thus not present in the nginx container, which serves static files)
- If a user is prompted to sign in because of the login_required() decorator, the user is then redirected to the relevant
  page after signing in.