added the entire "problems" feature
Users can now report problems that affect either a room or an equipment
(NB: There is no client or proper server-side validation that checks that a newly created problem does
not affect both a room and an equipment or neither a room nor an equipment. The backend will reject the
request but it's not hanled in a user-friendly way)

Users can also:
- delete the problems they reported
- add comments to already existing problems (provided that they are still open) and delete their
own comments.

Administrators can:
- delete any problem
- delete any comment on any problem
- mark a problem as either "SOLVED" or "CLOSED"