added user-friendly client-side validation when creating a problem
When creating a problem, it is no longer possible (on the client side) to submit the form
without selection neither a room nor a piece of equipment OR with selection both a room
and a piece of equipment.
Previously, submit such an invalid form would result in an HTTP 400 Bad Request.