fixed wrong timedelta being displayed on problems and comments
The days and years passed since the creation of problems and comments
were calculated bases on the timedelta.seconds field instead of the
timedelta.days field. The "seconds" field does not represent the total
amount of seconds of the timedelta but rather the amount of seconds
remaining after take into acount the number of "days".